+++ HISTORY.md
@@ -9,6 +9,7 @@
 ### 2025-02-26
 * threads: tweaked STRETCH COSTUME primitive to also accept a matrix of pixel data as input and to then also return one
 * updated dev version
+* extensions: device-orientation sensing extension, experimental
 
 ### 2025-02-25
 * gui: tweaked sound recording to - hopefully - support Safari, sigh